Add scripts to automate VM configuration
These scripts automate the installation and configuration of software
used in the course, specifically the following:

* Apache and PHP
* MySQL (and its PHP drivers)
* PostgreSQL and pgAdmin

The scripts target Ubuntu 18.04 (the current LTS release) and are
designed for students to use on a virtual machine (VM) to minimize the
number of issues encountered when installing the software,
particularly when compared to prior experiences trying to support this
process on Windows.

Scripting the installation and configuration of packages in a VM is
preferable to distributing a VM image for several reasons. First,
these scripts are much smaller than a VM image, requiring less storage
and network data transfer. Second, a VM image cannot be updated once
released. In comparison, these scripts are designed to be idempotent
-- i.e., they can be executed multiple times if there are updates
throughout the semester.

Restoring databases requires a separate file that is intentionally not
included in the repository. The rationale for this design decision is
twofold. First, it allows these scripts to be published publicly
without concern for the terms of distributing the databases (which
might be an issue for databases provided with a textbook). Second, the
US Air Force Academy currently imposes a 150-day limit on files shared
publicly from Microsoft OneDrive. Hence, each semester the links to
download the databases must be updated, which introduces considerable
(and unnecessary) churn in any repository that stores these links. The
only downside is that students must not only have a copy of the
scripts to create the databases but also the separate configuration
file.