Ignore blank lines in databases.yaml
A blank line at the end of databases.yaml causes the database
restoration to fail due to the missing URL (when splitting the
name of the database from the URL of the database dump). This
change skips blank lines in the file, which are typically added
accidentally (e.g., from copy-paste of the text).

Closes #8